+++ extensions/bonjour/src/service.ts
@@ -1,8 +1,28 @@
-import { Cache } from "@raycast/api";
+import { Cache, Color } from "@raycast/api";
 import Bonjour, { Service } from "bonjour-service";
+import { execSync } from "child_process";
 
 export const KEY = "services";
 export const cache = new Cache();
+
+declare global {
+  interface String {
+    get status(): Color;
+  }
+}
+
+Object.defineProperties(String.prototype, {
+  status: {
+    get: function () {
+      try {
+        execSync(`curl -I http://${this}`);
+        return Color.Green;
+      } catch {
+        return Color.Red;
+      }
+    },
+  },
+});
 
 export class HttpService extends Service {
   static set services(value: HttpService[]) {
@@ -27,15 +47,19 @@ export class HttpService extends Service {
 
   static fetch() {
     new Bonjour().find({ type: "http" }, (service: Service) => {
-      this.services = (this.services ?? []).concat(
-        this.services?.find((s) => s.name == service.name)
-          ? []
-          : ({
-              ...service,
-              origin:
-                `${service.type}://${service.host}:${service.port}`.toLowerCase(),
-            } as HttpService),
-      );
+      this.services = Object.values({
+        ...(this.services ?? []).reduce((result, service) => {
+          return {
+            ...result,
+            [service.name]: service,
+          };
+        }, {}),
+
+        [service.name]: {
+          ...service,
+          origin: `http://${service.host}:${service.port}`.toLowerCase(),
+        } as HttpService,
+      });
     });
   }