+++ extensions/bonjour/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Bonjour Changelog
 
+## [Improvements] - 2024-09-11
+
+- add status indicators
+
 ## [Bug Fixes] - 2024-08-26
 
 - fix cache value when no services present